Code for generating pseudoviolin plots of DGEs
These were DGEs determined by DESeq2 after limiting the RNA-Seq read counts for each NHP species to only genes with a one-to-one human ortholog. Since each species had different genes that fell into this one-to-one ortholog category, a corresponding human DGE profile was generated whereby the human RNA-Seq read counts were limited to these same genes prior to performing DESeq2. Once more, since reads were aligned to the human or the species-specific genome, this was performed twice -- once with the output of the human genome mapping method and a second time with the output of the species-specific genome mapping method. The results of the two mapping methods were subsequently limited to the same set of one-to-one ortholog genes for a given species.