Outputs of code
The plots shown in Fig S8 generated by the code. The beginning of the file name indicates genome mapping method ("HumanMapped" for mapping to the human genome and "SpeciesMapped" for mapping to the species-specific genome). "HumanRelated" indicates that these are the human DGE profiles when you limit the genes to those that have a one-to-one human ortholog on a species-by-species basis. "NHPs" indicates that these are the NHP DGE profiles after limiting to the same set of one-to-one human ortholog genes on a species-by-species basis. Mouse was not limited to one-to-one orthologs -- reads from mouse samples were mapped only to the mouse genome and then not limited further before performing DESeq2.